+++ include/roq/fbs/api.hpp
@@ -3336,7 +3336,8 @@ struct GatewaySettings FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
     VT_OMS_DOWNLOAD_HAS_ROUTING_ID = 16,
     VT_MBP_TICK_SIZE_MULTIPLIER = 18,
     VT_MBP_MIN_TRADE_VOL_MULTIPLIER = 20,
-    VT_OMS_REQUEST_ID_TYPE = 22
+    VT_OMS_REQUEST_ID_TYPE = 22,
+    VT_MBP_CHECKSUM = 24
   };
   uint64_t supports() const {
     return GetField<uint64_t>(VT_SUPPORTS, 0);
@@ -3365,6 +3366,9 @@ struct GatewaySettings FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
   roq::fbs::RequestIdType oms_request_id_type() const {
     return static_cast<roq::fbs::RequestIdType>(GetField<uint8_t>(VT_OMS_REQUEST_ID_TYPE, 0));
   }
+  bool mbp_checksum() const {
+    return GetField<uint8_t>(VT_MBP_CHECKSUM, 0) != 0;
+  }
   bool Verify(flatbuffers::Verifier &verifier) const {
     return VerifyTableStart(verifier) &&
            VerifyField<uint64_t>(verifier, VT_SUPPORTS, 8) &&
@@ -3376,6 +3380,7 @@ struct GatewaySettings FLATBUFFERS_FINAL_CLASS : private flatbuffers::Table {
            VerifyField<double>(verifier, VT_MBP_TICK_SIZE_MULTIPLIER, 8) &&
            VerifyField<double>(verifier, VT_MBP_MIN_TRADE_VOL_MULTIPLIER, 8) &&
            VerifyField<uint8_t>(verifier, VT_OMS_REQUEST_ID_TYPE, 1) &&
+           VerifyField<uint8_t>(verifier, VT_MBP_CHECKSUM, 1) &&
            verifier.EndTable();
   }
 };
@@ -3411,6 +3416,9 @@ struct GatewaySettingsBuilder {
   void add_oms_request_id_type(roq::fbs::RequestIdType oms_request_id_type) {
     fbb_.AddElement<uint8_t>(GatewaySettings::VT_OMS_REQUEST_ID_TYPE, static_cast<uint8_t>(oms_request_id_type), 0);
   }
+  void add_mbp_checksum(bool mbp_checksum) {
+    fbb_.AddElement<uint8_t>(GatewaySettings::VT_MBP_CHECKSUM, static_cast<uint8_t>(mbp_checksum), 0);
+  }
   explicit GatewaySettingsBuilder(flatbuffers::FlatBufferBuilder &_fbb)
         : fbb_(_fbb) {
     start_ = fbb_.StartTable();
@@ -3432,12 +3440,14 @@ inline flatbuffers::Offset<GatewaySettings> CreateGatewaySettings(
     bool oms_download_has_routing_id = false,
     double mbp_tick_size_multiplier = std::numeric_limits<double>::quiet_NaN(),
     double mbp_min_trade_vol_multiplier = std::numeric_limits<double>::quiet_NaN(),
-    roq::fbs::RequestIdType oms_request_id_type = roq::fbs::RequestIdType::Undefined) {
+    roq::fbs::RequestIdType oms_request_id_type = roq::fbs::RequestIdType::Undefined,
+    bool mbp_checksum = false) {
   GatewaySettingsBuilder builder_(_fbb);
   builder_.add_mbp_min_trade_vol_multiplier(mbp_min_trade_vol_multiplier);
   builder_.add_mbp_tick_size_multiplier(mbp_tick_size_multiplier);
   builder_.add_supports(supports);
   builder_.add_mbp_max_depth(mbp_max_depth);
+  builder_.add_mbp_checksum(mbp_checksum);
   builder_.add_oms_request_id_type(oms_request_id_type);
   builder_.add_oms_download_has_routing_id(oms_download_has_routing_id);
   builder_.add_oms_download_has_state(oms_download_has_state);